requirements.txt
requests
google-auth
google-auth-oauthlib
duckduckgo-search
questionary
rich
